travel-insurance/src/requirements.txt
Flask==2.2.2
numpy==1.23.1
pandas==1.4.2
pyodbc==4.0.34
PyYAML==6.0
Werkzeug==2.2.2
scikit-learn==1.1.2
waitress==2.1.2